chore: clean reporting module docstring
