+++ tests/runners/unified_runner.py
@@ -13,11 +13,11 @@ Provides all functionality in a single, maintainable system.
 import argparse
 import sys
 
-from src.utils.stability_improvements import stability_manager, safe_import
 from pathlib import Path
 from typing import Dict, List, Any, Optional
 
 from .base_runner import BaseTestRunner
+from .config import TEST_CATEGORIES
 
 # Color support
 try:
@@ -44,79 +44,8 @@ class UnifiedTestRunner(BaseTestRunner):
         """Initialize the unified test runner."""
         super().__init__(repo_root)
 
-        # Test categories from original run_tests.py
-        self.test_categories = {
-            "smoke": {
-                "description": "Smoke tests for basic functionality validation",
-                "marker": "smoke",
-                "timeout": 60,
-                "critical": True,
-                "directory": "smoke",
-            },
-            "unit": {
-                "description": "Unit tests for individual components",
-                "marker": "unit",
-                "timeout": 120,
-                "critical": True,
-                "directory": "unit",
-            },
-            "integration": {
-                "description": "Integration tests for component interaction",
-                "marker": "integration",
-                "timeout": 300,
-                "critical": False,
-                "directory": "integration",
-            },
-            "performance": {
-                "description": "Performance and load testing",
-                "marker": "performance",
-                "timeout": 600,
-                "critical": False,
-                "directory": "performance",
-            },
-            "security": {
-                "description": "Security and vulnerability testing",
-                "marker": "security",
-                "timeout": 180,
-                "critical": True,
-                "directory": "security",
-            },
-            "api": {
-                "description": "API endpoint testing",
-                "marker": "api",
-                "timeout": 240,
-                "critical": False,
-                "directory": "api",
-            },
-            "behavior": {
-                "description": "Behavior tree tests",
-                "marker": "behavior",
-                "timeout": 120,
-                "critical": False,
-                "directory": "behavior_trees",
-            },
-            "decision": {
-                "description": "Decision engine tests",
-                "marker": "decision",
-                "timeout": 120,
-                "critical": False,
-                "directory": "decision_engines",
-            },
-            "coordination": {
-                "description": "Multi-agent coordination tests",
-                "marker": "coordination",
-                "timeout": 180,
-                "critical": False,
-                "directory": "multi_agent",
-            },
-            "learning": {
-                "description": "Learning component tests",
-                "marker": "learning",
-                "timeout": 180,
-                "critical": False,
-                "directory": "learning",
-            },
-        }
+        # Test categories defined in a single source of truth
+        self.test_categories = TEST_CATEGORIES
 
     def print_banner(self):
         """Print the unified test runner banner."""
